Add composite directives examples
