+++ scala_akka/http/src/main/scala/highlevel/ExtractionDirectives.scala
@@ -5,34 +5,7 @@ import akka.http.scaladsl.model.{ContentTypes, HttpEntity}
 import akka.http.scaladsl.server.Directives._
 import common.HttpApp
 
-object Directives extends HttpApp {
-
-  /**
-   * Type #1: Filtering directives
-   */
-  private val simpleHttpMethodRoute = post {
-    complete(HttpEntity(ContentTypes.`text/plain(UTF-8)`, "POST"))
-  }
-
-  private val simplePathRoute = path("path") {
-    complete(HttpEntity(ContentTypes.`text/plain(UTF-8)`, "/path"))
-  }
-
-  private val complexPathRoute = path("complex" / "path") { // /complex/path
-    complete(HttpEntity(ContentTypes.`text/plain(UTF-8)`, "/complex/path"))
-  }
-
-  private val thisWouldBeUrlEncoded = path("url/encoded") {
-    complete(HttpEntity(ContentTypes.`text/plain(UTF-8)`, "/url%2Fpath"))
-  }
-
-  private val pathEndRoute = pathEndOrSingleSlash {
-    complete(HttpEntity(ContentTypes.`text/plain(UTF-8)`, "/"))
-  }
-
-  /**
-   * Type #2: Extraction directives
-   */
+object ExtractionDirectives extends HttpApp {
 
   private val pathExtractionRoute = path("api" / "items" / IntNumber) { (id: Int) =>
     complete(HttpEntity(ContentTypes.`text/plain(UTF-8)`, s"/api/items/$id"))
@@ -55,8 +28,7 @@ object Directives extends HttpApp {
     }
   }
 
-  private val router = simpleHttpMethodRoute ~ simplePathRoute ~ complexPathRoute ~ thisWouldBeUrlEncoded ~
-    pathEndRoute ~ pathExtractionRoute ~ pathMultiExtractionRoute ~ queryParamExtractionRoute ~ extractRequestRoute
+  private val router = pathExtractionRoute ~ pathMultiExtractionRoute ~ queryParamExtractionRoute ~ extractRequestRoute
 
   Http().newServerAt(host, defaultPort).bind(router)
